+++ .github/workflows/build-apps-mobile.yml
@@ -12,7 +12,7 @@ on:
       - 'packages/trpc/**'
       - 'packages/theme/**'
       - 'packages/mobile/ui/**'
-      - 'packages/mobile-forms/**'
+      - 'packages/mobile/forms/**'
       - 'packages/mobile-locales/**'
       - 'packages/mobile/storage/**'
       - 'packages/eslint-config-custom/**'
@@ -27,7 +27,7 @@ on:
       - 'packages/trpc/**'
       - 'packages/theme/**'
       - 'packages/mobile/ui/**'
-      - 'packages/mobile-forms/**'
+      - 'packages/mobile/forms/**'
       - 'packages/mobile-locales/**'
       - 'packages/mobile/storage/**'
       - 'packages/eslint-config-custom/**'
