+++ .ls-lint.yml
@@ -41,7 +41,7 @@ ignore:
   - packages/database/node_modules
   - packages/database/prisma/migrations
   - packages/eslint-config-custom/node_modules
-  - packages/mobile-forms/node_modules
+  - packages/mobile/forms/node_modules
   - packages/mobile-locales/node_modules
   - packages/mobile/storage/node_modules
   - packages/mobile/ui/node_modules
