+++ apps/mobile/package.json
@@ -11,8 +11,9 @@
     "lint:fix": "eslint --max-warnings 0 --fix --ext js,jsx,ts,tsx ."
   },
   "dependencies": {
-    "@acme/mobile-ui": "workspace:*",
+    "@acme/mobile-forms": "workspace:*",
     "@acme/mobile-storage": "workspace:*",
+    "@acme/mobile-ui": "workspace:*",
     "@expo/vector-icons": "^13.0.0",
     "@hookform/resolvers": "^2.9.10",
     "@react-native-async-storage/async-storage": "^1.17.11",
@@ -31,7 +32,6 @@
     "expo-status-bar": "^1.4.2",
     "i18next": "^22.0.5",
     "intl-pluralrules": "^1.3.1",
-    "mobile-forms": "workspace:*",
     "mobile-locales": "workspace:*",
     "react": "^18.1.0",
     "react-hook-form": "^7.39.3",
