+++ apps/mobile/screens/users/components/AddUserForm.tsx
@@ -7,7 +7,7 @@ import { addUserInputSchema } from 'trpc/schemas';
 import { Stack } from '@acme/mobile-ui/components/layouts';
 import { Button } from '@acme/mobile-ui/components/inputs';
 import { useSnackbarContext } from '@acme/mobile-ui/hooks/useSnackbarContext';
-import { TextField } from 'mobile-forms/components';
+import { TextField } from '@acme/mobile-forms/components';
 import { trpc } from '../../../utils/trpc';
 
 function AddUserForm() {
