+++ docs/architecture/apps/mobile.md
@@ -6,7 +6,7 @@ This app uses shared packages. For more info, read their docs:
 
 - [`trpc`](../packages/trpc.md)
 - [`mobile/ui`](../packages/mobile/ui.md)
-- [`mobile-forms`](../packages/mobile-forms.md)
+- [`mobile/forms`](../packages/mobile/forms.md)
 - [`mobile-locales`](../packages/mobile-locales.md)
 
 ## Software architecture
