+++ docs/architecture/packages/README.md
@@ -10,5 +10,4 @@ Configs are omitted, as they are pretty self-explanatory.
 - [Web Locales](./web-locales.md)
 - [Web Storage](./web-storage.md)
 - [Mobile libraries](./mobile/README.md)
-- [Mobile Forms](./mobile-forms.md)
 - [Mobile Locales](./mobile-locales.md)
