+++ docs/architecture/packages/mobile/README.md
@@ -1,4 +1,5 @@
 # Architecture of mobile packages
 
 - [UI](./ui.md)
+- [Forms](./forms.md)
 - [Storage](./storage.md)
