--- packages/mobile-forms/README.md
@@ -1,24 +1,0 @@
-# TreeView Mobile Forms library
-
-React Native form-specific component library and necessities for TreeView Mobile app.
-
-This package uses:
-
-- [`packages/mobile/ui`](../mobile/ui): UI Components
-
-## Setup
-
-```bash
-pnpm install
-```
-
-## Available scripts
-
-### Lint
-
-```bash
-pnpm lint
-
-# To fix lint
-pnpm lint:fix
-```
