+++ packages/mobile/forms/README.md
@@ -1,0 +1,24 @@
+# ACME Mobile Forms library
+
+React Native form-specific component library and necessities for ACME Mobile app.
+
+This package uses:
+
+- [`packages/mobile/ui`](../ui): UI Components
+
+## Setup
+
+```bash
+pnpm install
+```
+
+## Available scripts
+
+### Lint
+
+```bash
+pnpm lint
+
+# To fix lint
+pnpm lint:fix
+```
